Created a DAO and configured it as a service.
A DAO object gets injected with a SparQL client, an endpoint and a
cache object. It uses a set of stored SparQL queries to fetch data from
the triple store and caches the results. The data is accessible via
camelCased filenames of the queries. This is also used as a cache key.