Implemented preview page
Preview action uses twig to render the wiki text. Currently this is
still static dummy text. the rendered content is then converted using
pandoc to HTML. Both are then injected in the template and shown
separately.
The wiki text is editable and can be reconverted to generate an
up-to-date preview. This goes via an ajax call.